Add a manual result set key resolver because of API inconsistencies
The result set returned by the API in JSON format is nested in a key
that was assumed to be the same as the object name (camel case word),
but it turns out there not all objects adhere to this convention.

visitorActivity is returned as visitor_activity, but prospectAccount
stays prospectAccount.