Add warning message to XML handler, this is no longer supported
The XML response handler is not used by the current maintainer of this
library. Since there are unexpected issues with how the Pardot API
returns data, it does not make sense to invest any time in this.
If anyone actually wants to use this handler because they have to use
XML then get in touch with me.